backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gremlinpython==3.6.0
requests==2.31.0
python-dotenv==1.0.0
boto3==1.28.85
PyJWT==2.10.1
eth-account==0.13.7
